requirements.txt
# ─── Core ────────────────────────────────────────────
streamlit>=1.30.0
python-dotenv>=1.0.0

# ─── AI Providers ────────────────────────────────────
openai>=1.10.0
google-generativeai>=0.3.0

# ─── Data & Utilities ───────────────────────────────
pandas>=2.0.0

# ─── Testing ─────────────────────────────────────────
pytest>=7.0.0
